+++ src/greentest/greentest/patched_tests_setup.py
@@ -984,6 +984,22 @@ if PY37:
         'test_ssl.ThreadedTests.test_check_hostname_idn',
     ]
 
+    if APPVEYOR:
+        disabled_tests += [
+            # Raises "AssertionError: wait_threads() failed to cleanup
+            # 1 threads after 60.0 seconds (count: 3, old count: 2)". reason is
+            # unknown.
+            'test_socket.BasicTCPTest.testRecvFrom',
+            # Ditto.
+            'test_socket.BasicTCPTest.testRecv',
+            'test_socket.BasicTCPTest2.testDetach',
+            'test_socket.BasicTCPTest2.testFromFd',
+            'test_socket.BasicTCPTest2.testOverFlowRecvFrom',
+            'test_socket.BasicTCPTest2.testDup',
+            'test_socket.BasicTCPTest2.testSendAll',
+            'test_socket.NetworkConnectionAttributesTest.testTimeoutDefault',
+        ]
+
 # if 'signalfd' in os.environ.get('GEVENT_BACKEND', ''):
 #     # tests that don't interact well with signalfd
 #     disabled_tests.extend([